major refreshBlacklist + sessions refactor
refreshBlacklist is now completely removed and its functionality was added to sessions through the revoked_at field.
Several tests were added and fixed to test this refactor and other things

+++ IdentityX/internal/application/project/usecase.go
@@ -83,7 +83,6 @@ func (uc *UseCase) GetByID(ctx context.Context, projectID string) (*inbounds.Out
 	if err != nil {
 		return nil, err
 	}
-
 	pid, err := validation.RequireProjectID(span, &projectID)
 	if err != nil {
 		return nil, err
@@ -166,6 +165,17 @@ func (uc *UseCase) Update(ctx context.Context, in inbounds.ProjectServiceInput) 
 		return nil, err
 	}
 
+	isOwner, err := uc.projects.IsOwnerOf(ctx, *pid, principal.UserID)
+	if err != nil {
+		return nil, err
+	}
+
+	if !isOwner {
+		err = apierr.ErrUnauthorized.WithMsg("cannot update a project you don't own").WithID(apierr.ProjectNotOwnedByPrincipal)
+		apierr.RecordDomainError(span, err)
+		return nil, err
+	}
+
 	span.SetAttributes(attribute.String("project.id", *in.ProjectID))
 
 	newProject, err := uc.projects.GetByID(ctx, *pid, principal.UserID)
@@ -209,6 +219,18 @@ func (uc *UseCase) Delete(ctx context.Context, projectID string) error {
 	if err != nil {
 		return err
 	}
+
+	isOwner, err := uc.projects.IsOwnerOf(ctx, *pid, principal.UserID)
+	if err != nil {
+		return err
+	}
+
+	if !isOwner {
+		err = apierr.ErrUnauthorized.WithMsg("cannot delete a project you don't own").WithID(apierr.ProjectNotOwnedByPrincipal)
+		apierr.RecordDomainError(span, err)
+		return err
+	}
+
 	err = uc.projects.Delete(ctx, *pid, principal.UserID)
 	if err != nil {
 		return err
